Basic Website UI
Created an HTML file that contains a heading and a button. There's a bit of Javascript that allows some text to show up after the button is clicked. 

We also included a picture (trafficBackground.jpg) at some point which was meant to be used as a background picture, but we decided to comment it out for now and experiment with it later.